% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaply.R
\name{aapply}
\alias{aapply}
\title{aaply is like sapply but guaranteed to return a matrix}
\usage{
aapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object. Other objects (including classed objects) will be coerced by base::as.list.}

\item{FUN}{the function to be applied to each element of X. In the case of functions like +, %*%, the function name must be backquoted or quoted.}

\item{...}{option arguments to \code{FUN}}
}
\description{
aaply is like sapply but guaranteed to return a matrix
}
