% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_calc.R
\name{entropy_calc}
\alias{entropy_calc}
\title{Calculate entropy of list or vector}
\usage{
entropy_calc(x, naive = FALSE, hash = FALSE)
}
\arguments{
\item{x}{list or vector}

\item{naive}{logical use naive algorithm for entropy calculation}

\item{hash}{logical use hash}
}
\description{
Calculate entropy of list or vector
}
