% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtable.R
\name{mtable}
\alias{mtable}
\title{Create frequency table from multivariate data.}
\usage{
mtable(x)
}
\arguments{
\item{x}{matrix or dataframe}
}
\description{
How many times does a unique combination of variables occur?
}
\examples{
mtable(iris)
mtable(mtcars[,c(8:11)])
}
